+++ lib/src/ui/views/input/phone_input_field.dart
@@ -81,6 +81,7 @@ class PhoneInputWidgetView extends BaseStepView<QuestionStep> {
     if (!_hasRequestedFocus) {
       _hasRequestedFocus = true;
       WidgetsBinding.instance.addPostFrameCallback((_) {
+        if (isDisposed) return;
         _focusNode.requestFocus();
       });
     }